added update_user

+++ application.py
@@ -48,7 +48,7 @@ def update_pass():
 
         user = get_user(ref_id)
         if user["success"] is False:
-            return {"success": False, "timestamp": time.time()}
+            return {"success": False, "msg": "User not found!", "timestamp": time.time()}
 
         data = {"password": password, "timestamp": time.time()}
         data = json.dumps(data, indent=4)
@@ -68,7 +68,46 @@ def delete_user():
     try:
         ref_id = str(request.args['ref_id'])
 
+        user = get_user(ref_id)
+        if user["success"] is False:
+            return {"success": False, "msg": "User not found!", "timestamp": time.time()}
+
         resp = requests.delete(url=f"{url}/users/{ref_id}.json", headers=headers)
+
+        if resp.json() is None:
+            return {"success": False, "timestamp": time.time()}
+
+    except Exception as e:
+        return {"success": False, "msg": e, "timestamp": time.time()}
+
+    return {"success": True, "result": resp.reason, "timestamp": time.time()}
+
+@application.route('/update_user', methods=['POST'])
+def update_user():
+    try:
+        ref_id = str(request.args['ref_id'])
+        first_name = str(request.args['first_name'])
+        last_name = str(request.args['last_name'])
+        middle_initial = str(request.args['middle_initial'])
+        emergency_contact_number = str(request.args['emergency_contact_number'])
+        emergency_contact_person = str(request.args['emergency_contact_person'])
+
+        user = get_user(ref_id)
+        if user["success"] is False:
+            return {"success": False, "msg": "User not found!", "timestamp": time.time()}
+
+        data = {
+            "first_name": first_name, 
+            "last_name": last_name, 
+            "middle_initial": middle_initial, 
+            "emergency_contact_number": emergency_contact_number, 
+            "emergency_contact_person": emergency_contact_person, 
+            "timestamp": time.time()
+        }
+        
+        data = json.dumps(data, indent=4)
+
+        resp = requests.patch(url=f"{url}/users/{ref_id}.json", headers=headers, data=data)
 
         if resp.json() is None:
             return {"success": False, "timestamp": time.time()}
